requirements.txt
langchain==0.3.26
openai==1.91.0
langchain-community==0.3.26
httpx==0.28.1
python-docx==1.2.0
docx2txt==0.9
chromadb==1.0.13
tiktoken==0.9.0
python-dotenv==1.0.0
streamlit==1.39.0
